requirements.txt
# Core Flask dependencies
flask==3.0.2
flask-caching==2.1.0
flask-cors==4.0.0
gunicorn==21.2.0

# HTTP and API dependencies
requests==2.31.0

# Configuration and environment
python-dotenv==1.0.1

# Data processing and analysis
pandas==2.1.4
tabulate==0.9.0

# Time and date handling
pytz==2024.1
schedule==1.2.1

# Astronomical calculations for sunrise/sunset
astral>=3.2
timezonefinder>=6.2.0

# Optional: for enhanced geocoding (if you want to add this feature later)
# geopy>=2.3.0